Adapting to Coq PR #7257 (fix sensitivity of unification wrt alphabetic order). (#23)
In two occurrences of `econstructor` in this file, the unification
performed was dependent on the name of the variables involved in the
unification problem.

The unification problem was about solving

  ?s'@{a:=a; s':=s'; x:=a; x0:=s'} =? s'

and the variable coming first in ASCII order was chosen, namely
s'. Otherwise said, if x0 had be named, say l, then l would have been
chosen instead.

With PR #7257, the most recent name is chosen, independently of its
alphabetic value. Here, it means that x0 is now chosen.

We make the file compatible with Coq PR #7257 by changing the sequence
of tactics "econstructor. trivial" into "2:eassumption. constructor",
so that the expected instantiation of ?s' is done by "eassumption"
rather than by "econstructor". This is a backward compatible fix.